rename files per comment
